+++ src/main/java/bowling/BowlingApplication.java
@@ -1,5 +1,9 @@
 package bowling;
 
+import bowling.controller.BowlingController;
+import bowling.view.InputView;
+import bowling.view.ResultView;
+
 public class BowlingApplication {
 
     public static void main(String[] args) {
