+++ src/main/java/bowling/domain/FinalFrame.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 public class FinalFrame extends Frame {
