+++ src/main/java/bowling/domain/FrameState.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 public enum FrameState {
