+++ src/main/java/bowling/domain/PinNumber.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 public class PinNumber {
