+++ src/main/java/bowling/view/InputView.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.view;
 
 import java.util.Scanner;
