+++ src/test/java/bowling/FinalFrameTest.java
@@ -1,5 +1,7 @@
 package bowling;
 
+import bowling.domain.FinalFrame;
+import bowling.domain.PinNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
