+++ src/test/java/bowling/PinNumberTest.java
@@ -1,5 +1,6 @@
 package bowling;
 
+import bowling.domain.PinNumber;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
